+++ docs/attribute-curator/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 0.2.7 (unreleased)
 
+* Adds the `ncrad_samples` scope
 * Adds a `curated` tag to each curated file, and will skip curation on subsequent passes if the file has this tag
 
 ## 0.2.6